[fuzzing_event_engine] add necessary cast

+++ test/core/event_engine/fuzzing_event_engine/fuzzing_event_engine.cc
@@ -64,7 +64,7 @@ void FuzzingEventEngine::Restart(const fuzzing_event_engine::Actions& actions) {
   auto update_delay = [](std::map<intptr_t, Duration>* map,
                          fuzzing_event_engine::Delay delay, Duration max) {
     auto& value = (*map)[delay.id()];
-    if (delay.delay_us() > max.count() / GPR_NS_PER_US) {
+    if (delay.delay_us() > static_cast<uint64_t>(max.count() / GPR_NS_PER_US)) {
       value = max;
       return;
     }
